Build REX-Ray with Docker
This patch adds the ability to build REX-Ray with Docker for Linux or
Docker for Mac. To build REX-Ray with Docker, simply execute the
following command:

        $ make docker-build

The above command will copy the working source tree to a container
named "build-rexray". If a container with that name does not exist,
one is started using the "golang" image. If the container already
exists then it is not created again. This enables delta builds --
updating only what has changed.

Upon a successful build the resulting binary will be copied from the
container to $GOPATH/bin/rexray on the host OS.