TLS Support for ScaleIO GW 2.0.0.2
This patch adds support for ScaleIO GW 2.0.0.2. The GW update restricted
the available TLS ciphers to three, none of which are supported by
Golang. This patch updates the glide.yaml file to point to the GoScaleIO
branch "support/tls-sio-gw-2.0.0.2" which has a custom implementation of the
Golang "tls" package in order to support the TLS_RSA_WITH_AES_256_CBC_SHA256
cipher, one of the three ciphers supported by ScaleIO GW 2.0.0.2. This cipher
is also backwards compatible with older ScaleIO GW versions.