Storage Driver Sessions
This patch puts in place a new piece of core framework, a new interface
named "StorageDriverWithLogin" that defines a single function,
"Login(context.Context) (interface{}, error)". Types that implement the
existing "StorageDriver" interface may now also implement the
"StorageDriverWithLogin" interface by defining the new "Login" function.
Now part of the core workflow, the "Login" function is invoked for all
HTTP routes that consume a configured storage service and grant the
underlying storage driver implementation the ability to set a login
session object on the context of the current HTTP request.