InstanceID Fields
This patch adjusts the core framework's definition of the "InstanceID"
type as well as how "InstanceID" objects are consumed and used. The
"InstanceID" type now contains a field named "Fields" with a type of
"map[string]string". The "Fields" map may be leveraged by storage
executors to transport storage-platform specific information along
with the "InstanceID" object.

An "InstanceID" object still maintains the notion of metadata, but
the new "Fields" map is permanently associated with an "InstanceID"
object whereas metadata is set to nil once used by the remote
"InstanceInspect" API call. In fact, it is now permissible for
executor implementations to directly set an "InstanceID" object's
"ID" and/or "Fields" data, omitting the metadata entirely. If an
"InstanceID" object is without metadata the remote "InstanceInspect"
API call is not invoked in order to transform the metadata into the
actual instance identifier.

For executor implementations that are able to gather the complete
instance identifier from a client it is now more efficient for this
process to be used than relying on a remote API call to
"InstanceInspect" in order to transform an "InstanceID" object.