Handle Absent Instance ID for Region
This patch allows for the absence of an InstanceID (the transport for
the client's region information) when listing volumes for the case of
a controller client. If there is no instance ID present the storage
driver attempts to discover the region using the configuration file.
If there is still no region present the result is nondeterministic.