EBS Session Caching
This patch introduces EBS session caching. Now that EBS sessions are
scoped to HTTP requests and shuttled around with the Context object, it
becomes necessary to optimize their usage. Session objects are now
cached using a key that is the MD5 checksum of the concatenation of the
following string values that are identifying characteristics of an EBS
session: region, endpoint, and access key. Any unique combination of
those values (including if one is absent) will produce a unique session
object that is then cached for future use, available to requests with
the same identifying characteristics for establishing an EBS session.