Fix for Executor when Not Implemented
This patch should provide a means for a platform's executor code to
indicate that a function is not implemented on a given system. This way
the libStorage client may check for the "not implemented" error upon
initialization and not fail to start if one of storage platforms
available on a remote libStorage service is invalid for the current
client.