New CLI Revisions
This patch updates the volume CLI commands so that any command that
performs an action (all volume commands other than "ls" or "path") will
only perform the action when there are multiple results IFF a regular
expression was used to generate those results. If a provided string
matched more than one prefix of a volume's ID or name then the result
set is considered non-unique and will not be processed.

Additionally, a new flag -n|--dryRun has been introduced that will cause
several of the action-oriented volume commands to list the volumes that
will be acted upon, but not actually act upon them.