Fix for No Volumes Avai for some commmands
This patch updates REX-Ray to request a volume's attachments if a CLI
command is executed that depends upon a volume's status to determine its
candidacy for said operation. For example, only "available" volumes
(those not attached to another host) can be attached to a host. Only
volumes that are "attached" can be detached.

This patch handles issue 610
(https://github.com/codedellemc/rexray/issues/610) and part of issue 609
(https://github.com/codedellemc/rexray/issues/609).