Volume Attachment Filtering Logic
This patch updates the volume attachment filtering logic so that it can
return both attached and unattached volumes in the same query if both
bits are set in the mask. Additionally, if bit 2 is set it does not
restrict the volumes being returned to only those attached to the
provided instance unless the mask also includes bit 8.