Opt In Integration Driver
This patch makes it so that the integration driver is opt-in via a build
tag. Currently the in-tree tags available to introduce an integration
driver at build time are:

    * libstorage_integration_driver_docker

The tag "libstorage_integration_driver_docker" will build libStorage
with the Docker integration driver. Omitting this tag means libStorage
is not built with any integration driver and one must be supplied by the
embedding project.