Only set Vol AttachmentState if not set by driver
The framework can automatically set a volume's attachment state, but
a driver can also explicitly set it to "unknown" -- something that
only a driver can determine. That driver-set value was always
being overriden by the framework.

This patch makes it so that the framework only determines attachment
state if a driver has not already set a value.