Request devmap for integration volume unmount
When the integration driver does an unmount, it first does a volume
inspect on the requested volume ID. It was not requesting a devmap for
the attachments, but then expected the attachment.DeviceName to be
filled in, which it should not be.

Fix this by using the mask VolAttReqWithDevMapOnlyVolsAttachedToInstance
instead.