Show correct status after volume mount command
The displayed mount/attachment status was based on the volume info
retrieved *before* the mount occured. The integration driver is already
returning an updated volume struct, but it was being ignored. If you use
the updated value for output instead, you will see the correct status
displayed.

This mean status will show up as "attached" after a successful mount,
instead of the confusing "available".