Show correct status after volume unmount command
Use a returned volume object returned by the integration unmount
operation for output instead of volume object obtained at the beginning
of the unmount operation. The updated object has current/correct
attachment info, and therefore will display whether the volume is
actually availabe or still attached to another host.

Before this patch, the status would always display as "attached", which
is definitely not true when the unmount is successful.