Return volume from Docker integration Unmount()
This patch updates the IntegrationDriver types Unmount() method to
return a pointer to a Volume object. The unmount command is already
querying Libstorage to get an updated volume object after a detach has
occured, but it is not returning it.

Clients (like rexray) can use this returned value to display more up to
date status about the unmounted/detached volume without querying for it
a second time.