CLI Attachment Flags
This patch is in response, and intended to replace, PR #630. This patch
defines the CLI flags `--attached`, `--available`, and `--path`. If the
`--path` flag is provided, ex. `rexray volume ls --path`, then the
volume listing acts as if it were the now removed `volume path` command.
The `--path` flag supersedes the other two aforementioned flags which
themselves are mutually exclusive with the `--attached` flag taking
precedence over the `--available` flag. If the `--attached` flag is
provided then only volumes attached to the current instance are
returned. If the `--available` flag is specified then only volumes not
attached to any instance are returned.