Support for Creating Encrypted Volumes
This patch updates the `volume create` command to support the new flag
`--encrypted`. When specified, this flag sends an option to the
libStorage volume creation API that requests the creation of a new
storage volume using encryption. This feature does not guarantee
back-end encryption -- it's up to the libStorage storage driver and
underlying storage platform to support encryption.