Idempotent Commands
This patch adds the flag "-i|--idempotent" to the commands `volume
attach`, `volume detach`, `volume mount`, and `volume unmount`. Using
the new flag means that the command should always emit the same value
for a volume even after an operation has completed. For example, if
`volume attach` is used to attach a volume named "logs", without the
idempotent flag a subsequent `volume attach` operation on the "logs"
volume would result in an error indicating that no volumes are
available. This is because the above operations normally only act on
volumes that are in a state valid for the applicable operations. Because
making these commands idempotent adds overhead, the option is now
present but is something that has to be explicitly activated.