Classic, Client, Agent, & Controller Builds
This patch updates REX-Ray and the build scripts so that REX-Ray now
supports four build types: classic, client, agent, & controller.

        1. Classic      The REX-Ray you all know and love. This
                        binary includes a libStorage client,
                        Docker Volume Driver, and embedded
                        libStorage server.

        2. Client       A pure client binary sans Docker Volume
                        Driver service or embedded libStorage
                        server.

        3. Agent        A REX-Ray service that hosts the Docker
                        Volume Driver. This binary does not
                        include a client or embedded libStorage
                        server.

        4. Controller   A REX-Ray service that hosts an embedded
                        libStorage server. This binary does not
                        include a client or Docker Volume Driver.

The above REX-Ray types can be built with the following commands:

        1. Classic      Can be built with any of the following:

                        $ make
                        $ make deps && make build
                        $ make docker-build

        2. Client       Can be built with any of the following:

                        $ make deps && make build-client
                        $ make docker-build-client

        3. Agent        Can be built with any of the following:

                        $ make deps && make build-agent
                        $ make docker-build-agent

        4. Controller   Can be built with any of the following:

                        $ make deps && make build-controller
                        $ make docker-build-controller

CLEANING BETWEEN BUILDS
Unless the Docker option is used to build REX-Ray, it is necessary to
use `$ make clobber` to create a clean environment before building
another type. This is because build tags are used to control which
build is produced, and the build system is not able to consider build
tags when considering which targets need to be rebuilt. Thus if the
REX-Ray core package is already built in `client` mode and someone
attempts to build REX-Ray for `controller` mode, the core package
will not be rebuilt and may be incorrect for `controller` mode.

DOCKER BUILDS
However, the above issue is not a concern when building REX-Ray
with Docker since each build type uses a Docker container with the
prefix `rexray-build` and an eponymous suffix. For example, the
container used to build the REX-Ray client is named
`rexray-build-client` whereas the container used to build the
controller is named `rexray-build-controller`. The classic REX-Ray
build is still built with a container named `build-rexray`.

BUILD TYPE INFO
There are targets named `info`, `info-client`, `info-agent`, and
`info-controller` that can be used to obtain information about
each of the different build scenarios. There are also Docker
variants of these targets. For example, `docker-info`,
`docker-info-client`, etc.

INSTALLING
While nothing changes for the classic build type, the new build
types have some changes with regards to installation.

CLIENT INSTALLATION
The client build still supports the `install` and `uninstall`
commands so that the `rexray-client` binary can be
installed/uninstalled via the RPM/DEB package management systems,
but the actual commands do nothing.

AGENT INSTALLATION
The `rexray-agent` build is installed just as the classic mode,
with the exception that the agent's log and PID files are now named
`rexray-agent.log` and `rexray-agent.pid`. All other data directories
and behaviors are the same.

CONTROLLER INSTALLATION
The `rexray-controller` build is installed just as the classic mode,
with the exception that the controller's log and PID files are now
named `rexray-controller.log` and `rexray-controller.pid`. All other
data directories and behaviors are the same.