Fix installer for changed bintray HTML
When using the installers "list" functionality for a specifical release
channel (stable, staged, unstable), the script was no longer parsing the
possible versions correctly because the HTML has changed. There used to
be an "onclick" directive that is now now.

All versions are in an HREF tag that is surrounded by a <pre> block. So
install look for "pre".