Executor Supported Enhancements
This patch enhances the executor CLI to change the behavior of the
"supported" function. Internally, storage platforms will still implement
the same "Supported" function as previously, but now the executor CLI
returns more than simply a flag. The executor CLI will now return a
bitmask that corresponds to the functions supported by a storage
platform on the executing host. This provides a means to indicate
whether an executor supports mounting/unmounting for a given storage
platform. The update also centralizes the query for future additions to
the executor logic.

This patch also introduces the interface StorageExecutorWithMount with
two new functions, Mount and Unmount, but these functions are not yet a
part of the storage workflow.