Azure driver renaming and enhancements
This patchset renames the driver to `azureud`, to signify Azure
Unmanaged Disk driver. This set of enhancements builds off of previous
work, and completes the driver for inclusion into libStorage.

Use metadata server to detect azure instance

Don't return '.vhd' as part of volume name

Don't return OS disks

Proper attachment logic, remove instance filtering

improve login/init logic

Combine toTypeVolume and toTypesVolume

No need for separate functions here. This is in preparation for a call
to get VM details when getting attached device info. When we receive a
list of several volumes, and attachments.Devices() is true, we don't
want to have to query the VM status for every volume. By combining these
functions we can ensure that we only make one additional API call,
instead of making 1 call for every volume.

make executor localDevices return LUN number

The only way we have to match up an attached disk on a VM is via its LUN
number. When we query a disk, we get the name of the VM its attached to,
and that's it. When we query the VM, we get a list of data disks
attached to it, but not it's device path, only its LUN. Therefore, in
order to match up a volume with a device, we have to match on the LUN.

The easiest way to get the LUN is with the `lsscsi` utility. This
introduces a new, external dependency. It's technically possible to
parse through /sys/bus/scsi/devices/*/block/sd* to get this without an
external utility, but that is more effor than I am willing to put in at
this point.

This is the last hurdle needed to match up a volume with a device when
attachments.Devices() is requested.

Return attachment device info when requested

Remove IID restriction from VolumeCreate

VolumeCreate was requiring an IID, which is not necessary.