S3FS Fuse Driver Support
This patch introduces support for mounting S3FS buckets as Linux
file systems. This initial version of the S3FS storage driver does not
yet support volume (bucket) creation/removal.

This patch also updates the AWS SDK dependency that is shared by the EBS
and EFS storage drivers to v1.6.18, released on 2017/01/27.

This patch now includes support for force deleting a non-empty bucket.
All of the bucket's objects and their versions will be removed if
VolumeRemove is invoked with a truthy force value.