REX-Ray EBS Docker Plug-in
This patch introduces the first Docker 1.13 plug-in, REX-Ray for Amazon
EBS. The name of the Docker plug-in will be "rexray/ebs" and all future
Docker storage plug-ins based on REX-Ray will adhere to the same pattern
- "rexray/driver".

Due to Travis-CI's inability to take advantage of Docker when using
container builds coupled with non-container builds being so much slower,
the inclusion of support for automated Docker plug-in builds will rely
on enqueuing custom Travis-CI jobs at the end of a successful build.

All Docker plug-in-related material is found in the `.docker/plugins`
subdirectory. Several common files are in the `.docker/plugins`
directory and are copied or copied with sed into the
`.docker/plugins/driver` directory at build time.