Enhanced Docker Plug-in Builds
This patch provides the ability to create Docker plug-ins with multiple
drivers. Simply use `DRIVERS="ebs scaleio efs" make build-docker-plugin`
and voila. The one requirement is that a directory must exist for the
provided driver list with dash `-` characters in place of the spaces at
`.docker/plugins/DRIVERS`. For example, for the previous example the
directory would be `.docker/plugins/ebs-scaleio-efs`. Inside that
directory there should be a file named `config.json` which is the
plug-in's configuration file. The REX-Ray configuration file and its
multiple services for each driver is generated with the Makefile target.