ScaleIO Executor Enhancements
This patch enhances the ScaleIO executor so that the `drv_cfg` binary is
no longer hard-coded and can now be configured with the configuration
property `scaleio.drvCfg` which translates into the environment variable
`SCALEIO_DRVCFG`.

If the property `scaleio.guid` (env var `SCALEIO_GUID`) is set then the
`drv_cfg` binary is not used at all to discover the client's GUID.

This patch also updates the logic to discover local devices so the
`drv_cfg` is no longer necessary or required to provide this
information.