Service Names & Instance IDs
This patch updates the libStorage framework logic to store instance IDs
in maps with the service name as the key instead of the driver name.
This is to account for REX-Ray #685
(https://github.com/codedellemc/rexray/issues/685).

The instance ID format when marshaled as a string has been updated to
the following:

        DRIVER[:SERVICE]=ID,FIELDS,METADATA

This patch is backwards compatible with regards to the published
REST API. If an instance ID header does not include an optional
SERVICE component, the instance ID will be used for all services
with the specified driver type unless another instance ID header
with a specified service overrides it.