Ceph monitor port support (IPv4/IPv6)
Previous fix to support Ceph monitors with non standard ports caused a
new issue for monitors *without* ports. The call to net.SplitHostPort()
fails if the string does not contain a ":".

Rework the parsing of monitor addresses to account for the
presence/absence of ports, both for IPv4 and IPv6. As it turns out, The
Ceph config file also requires that IPv6 addresses (but not ports!) be
enclosed by brackets, which is the same requirement for
net.SplitHostPort(), and is something taken advantage of in this patch.

Tests added to cover possible combinations.