PID Chaos in Docker Plug-in
This patch fixes an issue where a stale PID file used by REX-Ray is
never properly removed and locks the program due to the Docker plug-in
model of fast-restarting a plug-in if misconfigured combined with the
entry point always using PID 1.

Additionally, this patch updates the Dockerfiles so that the entrypoint
is now defined with default arguments that can be overridden via the CLI
for test and development purposes.