Use local device map for attachment device info
This patch changes the dobs driver to use the client-provided local
device map when looking up device info. Previously the driver was
assuming that it had local access to the device, but the required info
is actually provided by the client. This allows the the driver to work
in server/client mode.