Parse Request Opts Config Option
This patch defines a new configuration property
"libstorage.server.parseRequestOpts". When this property is set to true,
incoming POST requests will be parsed for an opts field, and key/value
pairs in the opts field's map will be promoted to the request proper.

For example, the following volume creation POST request bodies are now
equivalent when the above property is enabled:

        {
            "name":      "newVolume",
            "encrypted": true
        }

        {
            "name": "newVolume",
            "opts": {
                "encrypted": true
            }
        }