Always return Not Found when volume is not there
From VolumeInspect, it is not allowed to return nil, nil if the
referenced volume is not found. However, several drivers do so, and this
can lead to a nil dereference.

Additionally, for API calls where a specific volumeID is given, if that
volume does not exist, we are supposed to return a not found error.
THere is a helper function to do this -- use it.