Add minimum vol size, reject request too small
For each driver that uses the opt.Size field to create a volume with the
requested size, set a minimum volume size. If the incoming request does
not set a size, default to that minimum size. If the incoming request
does set a size, make sure that it is >= to the minimum size. If it is
not, return an error.