Move dobs client to storage package
The dobs client code was in the utils package, which is imported by the
executor. The dobs client also imports the "godo" package, which is a
non standard lib package that we try to keep out of the executors. The
executor was not using the dobs client at all, so instead move that code
into the actual dobs storage driver package.