Known Host Fix
This patch changes the known host logic for TLS connections so that the
saved host token of a known host is the host name for the attempted
connection's endpoint.

This patch also introduces a new error type, ErrKnownHostConflict. This
error can occur when a known host's fingerprint matches a remote peer's
fingerprint, but the known host's host name does not match the targeted
connection's host name.