Correctly handle volume not exist during create
The Isilon driver was checking if a volume with the same name already
exists before attempting to create. A recent change made it such that an
error was returned when querying for the volume, rather than a nil
error, causing the driver to always return the error.

Instead, check if the error is a NotFoundError. If it is *not*, return
it the error. If it *is*, that's expected and move on.