Add framework support for InspectVolumeByName
This commit introduces suport for InspectVolumeByName to the libStorage
client. When invoked, the libStorage server will execute an optimized
inspection of a single volume based on its name for drivers that support
it. This is an enhancement over the existing behavior of the client
requesting all volumes, then matching on the name to get the volumeID
and calling InspectVolume on that ID.

This patch does not update any drivers to take advantage of the new
support, and therefore should not change any existing behavior.