Module-Scoped Host Definitions
This patch fixes #827 where a libStorage host may not be defined at the
root of a configuration but instead scoped to a module. With this patch
if host detection fails (root configuration / running, local host) when
initializing the module subsystem a warning is logged and the modules
themselves must explicitly define the host to which to connect.

This patch does *not* support starting a new host from a module-scoped
host definition. The module-scoped host properties are for overriding
the host to which to connect only.