No longer report invalid local devices
The Cinder driver cannot completely fill out the local device struct
from the executor because it cannot lookup the source volume ID for
attached volumes. Instead, it was reporting an empty volume ID, which is
invalid and was subsequently removed by the libStorage client when
parsing output from the xcli. This just resulted in wasted cycles and
extra log messages.