Add dobs support for VolumeInspectByName
This patch add the VolumeInspectByName function to the dobs driver,
allowing a caller to more efficiently retrieve a single volume when
doing a lookup by name instead of by ID.

When looking up a volume by name instead of ID, the DO API requires a
region, so the way a region was looked up was changed to have
commonality between VolumeCreate() and VolumeInspectByName().
VolumeCreate() retains the additional capability to read a region from
the AvailabilityZone create option.

The order of precedence for region is:
1. az op (create only)
2. IID header field
3. driver config