EFS SecurityGroup Warning
This patch fixes #511 by sending an instance's full SG list to the
remote service where the EFS storage driver examines the server-side SG
list with the client-side list. If any of the server-side SGs are not
present in the client-side list a warning will be logged in order to
assist with debugging.