Upload Build Info
This patch updates CI builds so that upon success or failure a README is
pushed to a gist that provides instructions on how to duplicate the
build exactly using Docker. This process should produce a binary with
the exact same checksum as the one built on the Travis-CI build server.