Fix for Docker Module
This patch removes the Go build tag required to include the Docker
module. REX-Ray standard and agent binaries should always include this
module.

Prior to this change there was a bug where a vestigial tag
"libstorage_integration_driver_linux" was expected to exist in order to
build support for the Docker module. Because the build process does not
provide that tag any more, binaries produced with `go build` would omit
the Docker module.