Implement Node/Controller Publish/Unpublish
This fills out the logic for Controller Publish/Unpublish, and Node
Publish/Unpublish. The controller Publish uses the libStorage
VolumeAttach() logic, which merely attaches the block device to the
given node and then returns the device token in the PublishVolumeInfo.
This is important as that token needs to be provided for Node Publish
operations.

Controller Unpublish does the reverse, and detaches the volume from the
node.

Node publish uses the existing integration driver's Mount logic, which
will mount the device in the standard /var/lib/rexray/volumes/...
location, or wherever it is configured to do so. It also adds the /data
directory into the volume when it is mounted. The /data dir is
bind-mounted to the taret.  Node publish also will bind-mount the
underlying device, specificied by the PublishVolumeInfo token, directly
to a file specified by target if BlockVolume is used instead.

Node Unpublish removes the bind mount, and the private mount in
/var/lib/rexray if the volume is not mounted anywhere else.