Account for root path in libStorage volume
This patch makes sure that when looking for an existing mount, we
account for the root path of libStorage volumes. The root path is
typically `/data`, and it is *that* path that is bind-mounted to the
target.

We also use localDevices to get the actual source device to look for in
the mount table when we have access to the device token that is returned
from the Conreturned from the ControllerPublishVolume. It also has the
 side-effect of eliminating a call to the libStorage server.