Update idemp provider to handle block and mnt vols
This patch updates the idempotency provider to handle both block and
mount volume types, as they need to be handled a bit differently. This
information is not always given directly in the CSI request, so right
now it is inferred by the target path. If the path is a dir, we assume
mount, if the path is a file, we assume block.

We also simplify the logic of looking for mounts (to see if published or
not) by only matching the source device and the target, rather than also
caring about the intermediate "private" mount in between. We can do this
because the interceptor has been been updated previously to catch any
errors returned from the CSI bridge, and if there are any "dangling"
mounts, we trust our bridge to return an error, which should cause a
retry.

We also updated the CSI bridge to behave directly when the idempotency
interceptor and provider are not in play -- the logic to look for
existing mounts was not in place, but is now.