CLI Error Handling
This patch updates CLI error handling at the request of @cduchesne.
Instead of early, startup errors resulting in panics, the errors now are
emitted to STDERR. For example:

        $ ./rexray start -s vfs1
        error: service startup failed: start libStorage failed: \
          invalid driver name: vfs1

        $ ./rexray start -s vfs -h tcp://2
        error: service startup failed: agent: mod init failed: \
          Get http://2/services: dial tcp: address 2: missing port \
          in address

Errors that are not related to immediate startup routines may still
result in panics, but any "early" errors should be emitted to STDERR
without a stack trace.

Additionally, the REX-Ray ASCII logo and version are not emitted to
STDOUT until *after* all services have started. If any service, such
as the Agent or Controller, return an error during startup, the error
is printed to STDERR and the program exists before the ASCII logo and
version are emitted.